Invoke algoritms inside dedicated temporary directory
Given analysis is invoked in dedicated directory for particular
repository. For instance if project "foobar" was analysed all
files will be stored in /tmp/code-compass-foobar like so:

$ tree /tmp/code-compass-foobar
/tmp/code-compass-foobar
├── cloc.csv
├── csv_as_enclosure_json.py
├── d3
│   └── d3.min.js
├── gitreport.log
├── hotspot_proto.json
├── revisions.csv
├── script.js
├── style.css
└── zoomable.html

This change makes it easier to write algorithm steps since they can be
written as-if they were invoked inside that dedicated temporary
directory. It also opens up possibility to have multiple analyses
without them overriding each other - not in the scope of this change
but it could easily be achieved with c/temp-dir depending not only on
repository but for instance analysis name, period etc.