simplify caching of coupling data
This modifies the key for the hashmap to include the revision hash.
This makes some of the code introduced for #33 redundant. I removed
also the backend check because that concern is cross cutting and I
would need a significant refactor of code if anybody ever asks for
support for other VCS.

Closes #38.